complete project be

+++ src/controllers/document.controller.js
@@ -12,10 +12,16 @@ const DocumentControllers = {
             document_content,
             store_Id,
             document_score,
-            format_Id
+            format_Id,
+            document_author,
+            document_year,
         } = req.body
 
-        if (!document_title || !store_Id || isNaN(+document_score) || !format_Id || !req.file) {
+        if (!document_title || !store_Id
+            || isNaN(+document_score)
+            || !format_Id || !req.file
+            || !document_author || isNaN(+document_year)
+        ) {
             return res.errorValid()
         }
 
@@ -39,9 +45,17 @@ const DocumentControllers = {
                 )
             }
 
+            // Lấy dung lượng file gốc
+            const fileStats = fs.statSync(filePath);
+            const fileSizeInBytes = fileStats.size; // Dung lượng file tính bằng bytes
+            const fileSizeInKB = (fileSizeInBytes / 1024).toFixed(2); // Dung lượng tính bằng KB (2 chữ số thập phân)
+
             // Đọc file PDF gốc
             const existingPdfBytes = fs.readFileSync(filePath);
             const pdfDoc = await PDFDocument.load(existingPdfBytes);
+
+            // Lấy tổng số trang của file gốc
+            const totalPages = pdfDoc.getPageCount();
 
             // Đọc file logo
             const logoBytes = fs.readFileSync(logoPath);
@@ -52,6 +66,7 @@ const DocumentControllers = {
 
             // Lấy tất cả các trang của PDF
             const pages = pdfDoc.getPages();
+
             for (const page of pages) {
                 const { width, height } = page.getSize();
 
@@ -75,6 +90,34 @@ const DocumentControllers = {
             const modifiedFilePath = path.join(uploadFolder, `modified_${req.file.filename}`);
             const modifiedFileSave = path.join('uploads', `modified_${req.file.filename}`);
             fs.writeFileSync(modifiedFilePath, modifiedPdfBytes);
+
+            // Tạo PDF chỉ với 10% số trang hoặc tối đa 15 trang
+            const shortPdfDoc = await PDFDocument.create();
+            const numShortPages = Math.min(Math.max(1, Math.ceil(totalPages * 0.1)), 15); // Tính số trang
+            for (let i = 0; i < numShortPages; i++) {
+                const [copiedPage] = await shortPdfDoc.copyPages(pdfDoc, [i]);
+                shortPdfDoc.addPage(copiedPage);
+
+                const { width, height } = copiedPage.getSize();
+                const scaledLogoWidth = width / 4;
+                const scaledLogoHeight = (logoHeight / logoWidth) * scaledLogoWidth;
+
+                copiedPage.drawImage(logoImage, {
+                    x: width - scaledLogoWidth - 50,
+                    y: height - scaledLogoHeight - 50,
+                    width: scaledLogoWidth,
+                    height: scaledLogoHeight,
+                    opacity: 0.2
+                });
+            }
+
+            // Lưu PDF có 10% số trang hoặc tối đa 15 trang
+            const shortPdfBytes = await shortPdfDoc.save();
+            const shortFilePath = path.join(uploadFolder, `short_${req.file.filename}`);
+            const shortFileSave = path.join('uploads', `short_${req.file.filename}`);
+            fs.writeFileSync(shortFilePath, shortPdfBytes);
+
+            // Xóa file gốc
             fs.unlinkSync(filePath);
 
             const newDocument = await DocumentServices.create(
@@ -84,7 +127,12 @@ const DocumentControllers = {
                     store_Id,
                     document_score: +document_score,
                     format_Id,
-                    document_url: modifiedFileSave
+                    document_url: modifiedFileSave,
+                    document_sub: shortFileSave,
+                    document_capacity: +fileSizeInKB, // Dung lượng file
+                    document_page: totalPages, // Số trang file,
+                    document_author,
+                    document_year: +document_year
                 }
             )
 
@@ -99,7 +147,6 @@ const DocumentControllers = {
                 'Thêm mới tài liệu thất bại!'
             )
         } catch (err) {
-            console.log(err)
             return res.errorServer()
         }
     },
@@ -108,25 +155,33 @@ const DocumentControllers = {
             document_title,
             document_content,
             store_Id,
-            document_score
+            document_score,
+            document_author,
+            document_year
         } = req.body
 
         const { id } = req.params
 
-        if (!document_title || !id || isNaN(+document_score) || !store_Id) {
+        if (!document_title || !id || isNaN(+document_score) || !store_Id || !document_author || isNaN(+document_year)) {
             return res.errorValid()
         }
 
         const transaction = await db.sequelize.transaction()
         try {
 
+            const data = {
+                document_title,
+                store_Id,
+                document_score: +document_score,
+                document_author,
+                document_year: +document_year
+            }
+
+            if (document_content) {
+                data.document_content = document_content
+            }
             const document = await DocumentServices.update(
-                {
-                    document_title,
-                    document_content: document_content ?? '',
-                    store_Id,
-                    document_score: +document_score,
-                },
+                data,
                 id,
                 transaction
             )
@@ -176,11 +231,11 @@ const DocumentControllers = {
         }
     },
     async getAll(req, res) {
-        const { page, limit, score, format, store, title } = req.query
+        const { page, limit, score, format, store, title, auth, year, id, order } = req.query
 
         try {
             const documents = await DocumentServices.getAll({
-                page, limit, score, format, store, title
+                page, limit, score, format, store, title, auth, year, id, order
             })
 
             if (documents) {
@@ -225,8 +280,10 @@ const DocumentControllers = {
 
             if (data) {
                 const filePath = path.join(__dirname, '../' + data.document_url);
-                if (typeof filePath === 'string' && fs.existsSync(filePath)) {
+                const filePathSub = path.join(__dirname, '../' + data.document_sub);
+                if (typeof filePath === 'string' && fs.existsSync(filePath) && typeof filePathSub === 'string' && fs.existsSync(filePathSub)) {
                     fs.unlinkSync(filePath);
+                    fs.unlinkSync(filePathSub);
                     const document = await DocumentServices.delete(id)
 
                     if (document) {
